requirements.txt
aiohttp==3.9.5
aioresponses==0.7.6
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
backoff==2.2.1
certifi==2024.7.4
charset-normalizer==3.3.2
colorama==0.4.6
contourpy==1.2.1
cvlib==0.2.7
cycler==0.12.1
dataclasses-json==0.6.7
defusedxml==0.7.1
filelock==3.15.4
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
idna==3.7
imageio==2.34.2
importlib_resources==6.4.0
imutils==0.5.4
inference-sdk==0.14.1
intel-openmp==2021.4.0
Jinja2==3.1.4
kiwisolver==1.4.5
MarkupSafe==2.1.5
marshmallow==3.21.3
matplotlib==3.9.1
mkl==2021.4.0
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
networkx==3.2.1
numpy==1.26.4
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
packaging==24.1
pandas==2.2.2
pillow==10.4.0
progressbar==2.5
psutil==6.0.0
py-cpuinfo==9.0.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
scipy==1.13.1
seaborn==0.13.2
six==1.16.0
supervision==0.22.0
sympy==1.13.0
tbb==2021.13.0
torch==2.3.1+cpu
torchvision==0.18.1
tqdm==4.66.4
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
ultralytics==8.2.58
ultralytics-thop==2.0.0
urllib3==2.2.2
yarl==1.9.4
zipp==3.19.2
